6-size.c

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ int main(void)
 	long long int d;
 	float e;
 
-	Printf("size of a char: %lu byte(s)\n", sizeof(a));
-	Printf("size of an int: %lu byte(s)\n", sizeof(b));
-	Printf("size of a long int: %lu byte(s)", sizeof(c));
-	Print("size of a long long int: %lu byte(s)\n", sizeof(d));
-	Print("size of a float: %lu byte(s)\n", sizeof(e));
+	Printf("size of a char: %lu byte(s)\n", (unsigned long)sizeof(a));
+	Printf("size of an int: %lu byte(s)\n", (unsigned long)sizeof(b));
+	Printf("size of a long int: %lu byte(s)", (unsigned long)sizeof(c));
+	Print("size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
+	Print("size of a float: %lu byte(s)\n", (unsigned long)sizeof(e));
 	return (0);
 }
